Change code to use gModes[...] object.
Game mode logic is now contained in the gModes associative array. This is
simply an object that contains any of the following methods:
  * start
  * reset
  * draw
  * loop(delta)
  * tap(x, y)

The current mode is determined by the last focused mode element by taking
its id of "mode-...", stripping off the initial "mode-" string, and using
that to reference the associative array.

Additionally, the global "lVariableName" syntax has been replaced with
"gVariableName", to better represent the intent. "lVariableName" is now
used as local variables within modes.